requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.21
botocore==1.23.21
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
fastapi==0.70.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
jmespath==0.10.0
lxml==4.6.4
protobuf==3.19.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.8.2
PyJWT==2.3.0
pymongo==4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
random-object-id==2.0.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
ua-parser==0.10.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
